refactor:コンポーネントディレクトリ整理 #9
