+++ resources/ts/Components/Book/FavoriteIcon.tsx
