+++ resources/ts/Components/Icon/ApplicationLogo.tsx
