+++ resources/ts/Components/Memo/MemoCard.tsx
