+++ resources/ts/Pages/Auth/ForgotPassword.tsx
@@ -1,4 +1,4 @@
-import InputError from '@/Components/InputError';
+import InputError from '@/Components/ui/InputError';
 import { Button } from '@/Components/ui/button';
 import { Input } from '@/Components/ui/input';
 import DefaultLayout from '@/Layouts/DefaultLayout';
