+++ resources/ts/Pages/FavoriteBookList.tsx
@@ -1,5 +1,5 @@
 import BookCard from '@/Components/BookCard';
-import BookShelfDescription from '@/Components/BookShelfDescription';
+import BookShelfDescription from '@/Components/BookShelf/BookShelfDescription';
 import DefaultLayout from '@/Layouts/DefaultLayout';
 import { Head } from '@inertiajs/react';
