+++ resources/ts/Pages/Profile/Partials/DeleteUserForm.tsx
@@ -1,4 +1,4 @@
-import InputError from '@/Components/InputError';
+import InputError from '@/Components/ui/InputError';
 import { Button } from '@/Components/ui/button';
 import {
   DialogDescription,
